Move The EventDescriptor to Common from Platform-Core
