+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/tipset/ChildlessEventTracker.java
@@ -17,7 +17,7 @@
 package com.swirlds.platform.event.tipset;
 
 import com.swirlds.common.system.NodeId;
-import com.swirlds.platform.event.EventDescriptor;
+import com.swirlds.common.system.events.EventDescriptor;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.ArrayList;
 import java.util.HashMap;
